Do not drop privileges
Running as root is the simple way to access /var/run/docker.sock from
within the container.

Adding the unprivileged user to the `docker` group is the recommended
way to grant access to /var/run/docker.sock. However, at build time
the group ID is unknown, unless we are building the image on the
target system.

In the future, we may consider adding `buildbot` to the `docker` group
from within the entrypoint, and dropping privileges before `exec`.