Construct buildbotURL from DOCKER_HOST
The default of "localhost" works fine when running in a single Docker
installation on a developer machine.

Otherwise, we need to pass the IP address or hostname at which we
access the web interface.

When using docker-machine, the IP address is contained in the
environment variable DOCKER_HOST, so we can use that. For all other
cases, we should also provide an environment variable to specify the
buildbot master host explicitly (TBD).