Bind-mount configuration to /etc/buildbot/master.cfg
Put a symlink in buildbot's basedir, to unconfuse it. Bind-mounting
the config there directly does not work because `chown` in the
entrypoint fails on a read-only file system.

Do not create the symlink if it already exists. The symlink is created
in the volume, so it will persist between containers.