Add BUILDBOT_URL to service environment
Configure the external URL at which the web interface is accessed
using the BUILDBOT_URL environment variable. This corresponds to the
`buildbotURL` entry in `BuildmasterConfig`.

If this environment variable is set in the shell from which the stack
is deployed, it will be exported to the service environment.

If BUILDBOT_URL is unset but DOCKER_HOST is set, its value is used to
provide a default for BUILDBOT_URL. Otherwise, `buildbotURL` defaults
to `http://localhost:8010`.

An incorrect buildbot URL results in errors in the web interface,
because requests to the API violate the Same Origin Policy. An example
would be the following message when triggering a build:

    "invalid origin"