Create master.cfg symlink at build-time
Create the symlink to /etc/buildbot/master.cfg when the image is
built, rather than in the entrypoint.

There is no need to delay symlink creation to runtime. Like this, we
don't need to check for existence of the symlink. Apart from being
simpler, this solution also gives us a working `buildbot` command when
the image is run without the entrypoint script.